Moved cognito provider into custom providers
